fix: moved Thumbnail component to Video.tsx
